import of ConfigurationEntry moved to "data" package
