+++ src/main/java/de/mme/cfm/configurations/Configuration.java
@@ -1,6 +1,6 @@
 package de.mme.cfm.configurations;
 
-import de.mme.cfm.ConfigurationEntry;
+import de.mme.cfm.data.ConfigurationEntry;
 
 public interface Configuration {
     ConfigurationEntry getEntry(String name);
