+++ src/test/java/de/mme/cfm/testing/tests/uniqueConfiguration/Test_UniqueConfiguration_setEntry_Configurationentry.java
@@ -1,6 +1,6 @@
 package de.mme.cfm.testing.tests.uniqueConfiguration;
 
-import de.mme.cfm.ConfigurationEntry;
+import de.mme.cfm.data.ConfigurationEntry;
 import de.mme.cfm.configurations.Configuration;
 import de.mme.cfm.configurations.UniqueConfiguration;
 import org.junit.jupiter.api.Assertions;
